requirements.txt
gtts==2.2.3
googletrans==4.0.0-rc1
streamlit==1.8.0
sounddevice==0.4.1
soundfile==0.10.3.post1
transformers==4.9.2
pillow==8.3.1
pdf2image==1.16.0
pytesseract==0.3.8
